refactor: 숙소 필터를 담당하는 클래스가 기존 Repository를 상속하도록 변경 #18
