+++ BE/airdnb/src/main/java/com/team01/airdnb/accommadation/AccommodationRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccommodationRepository extends JpaRepository<Accommodation, Long> {
-
+public interface AccommodationRepository extends JpaRepository<Accommodation, Long>, AccommodationFilterRepository {
 
 }
